🔥 chat 파일 삭제 및 이동
