+++ src/components/common/Chat/ChatInput.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 
 import { ChatBar, Button } from "./ChatInputStyle";
-import RoundInput from "../../components/common/Input/RoundInput";
-import PlusIcon from "../../assets/icons/icon-plus.svg";
+import RoundInput from "../Input/RoundInput";
+import PlusIcon from "../../../assets/icons/icon-plus.svg";
 
 function ChatInput() {
   const [message, setMessage] = useState("");
